API References

YaoCompiler.AdjointOperation — Type

AdjointOperation{P} <: Routine

An AdjointOperation is the adjoint of another Routine.

YaoCompiler.AnyReg — Type

AnyReg <: AbstractRegister{1}

A place holder for registers when compilation is not register specific.

YaoCompiler.GenericRoutine — Type

GenericRoutine{name} <: Function

A GenericRoutine cannot be directly execute on a quantum device. It is a Julia Function that returns Operation, and Operation can be executed on quanutm device.

Note

An instance of GenericRoutine should be treated like Function.

YaoCompiler.IntrinsicRoutine — Type

abstract type IntrinsicRoutine <: Routine end

IntrinsicRoutine are the routine that can be executed by the compile target that is pre-defined in the compiler.

YaoCompiler.NativeJuliaTarget — Type

NativeJuliaTarget <: YaoCompileTarget

This target assumes one compiles to native Julia, thus the measurement type will be directly inferred from native intrinsics.

YaoCompiler.Operation — Type

Operation{P, Args} <: Routine

An Operation is a user defined composite routine that can be called in other Operation or execute on target device.

YaoCompiler.Routine — Type

abstract type Routine end

This defines operations that one can execute on a quantum device.

YaoCompiler.routine_stub — Function

routine_stub(routine, args...)

A function to store the raw CodeInfo of defined routines to support multiple dispatch. We don't directly define main gate, ctrl etc. because we can't support multiple dispatch on them directly.

YaoCompiler.target_measure_result_type — Method

target_measure_result_type(target::YaoCompileTarget[, interp::YaoInterpreter, f, fargs, argtypes, sv, max_methods])

Hook for handling different measurement result type. Default type is MeasureResult.

YaoCompiler.@barrier — Macro

@barrier locs

Syntax sugar for barrier(locs).

YaoCompiler.@ctrl — Macro

@ctrl <ctrl_locs> <locs> => <gate>

Syntax sugar for apply(gate, locs, ctrl_locs), must be used inside @device. See also @device.

YaoCompiler.@device — Macro

@device <function def>

Annotate a Julia function as YaoLang device kernel.

YaoCompiler.@gate — Macro

@gate <locs> => <gate>

Syntax sugar for apply(gate, locs), must be used inside @device. See also @device.

Tips

You don't have to write @gate in most cases inside @device. But in case there is ambuigity, you can annotate the expression with @gate explicitly.

YaoCompiler.@measure — Macro

@measure locs

Syntax sugar for measure(locs).