API References

YaoCompiler.AnyReg — Type

AnyReg <: AbstractRegister{1}

A place holder for registers when compilation is not register specific.

YaoCompiler.NativeJuliaTarget — Type

NativeJuliaTarget <: YaoCompileTarget

This target assumes one compiles to native Julia, thus the measurement type will be directly inferred from native intrinsics.

YaoCompiler.routine_stub — Function

routine_stub(routine, args...)

A function to store the raw CodeInfo of defined routines to support multiple dispatch. We don't directly define main gate, ctrl etc. because we can't support multiple dispatch on them directly.

YaoCompiler.target_measure_result_type — Method

target_measure_result_type(target::YaoCompileTarget[, interp::YaoInterpreter, f, fargs, argtypes, sv, max_methods])

Hook for handling different measurement result type. Default type is MeasureResult.

YaoCompiler.@barrier — Macro

@barrier locs

Syntax sugar for barrier(locs).

YaoCompiler.@ctrl — Macro

@ctrl <ctrl_locs> <locs> => <gate>

Syntax sugar for apply(gate, locs, ctrl_locs), must be used inside @device. See also @device.

YaoCompiler.@device — Macro

@device <function def>

Annotate a Julia function as YaoLang device kernel.

YaoCompiler.@gate — Macro

@gate <locs> => <gate>

Syntax sugar for apply(gate, locs), must be used inside @device. See also @device.

Tips

You don't have to write @gate in most cases inside @device. But in case there is ambuigity, you can annotate the expression with @gate explicitly.

YaoCompiler.@measure — Macro

@measure locs

Syntax sugar for measure(locs).